Separate Image from PortableImage
The image coder only needs the w/h and data pointer, not any
other PPM related members.

Additionally, make Image a template class, to allow setting the
Sample datatype (for 8 bits vs. 16 bits per sample), and the
number of samples per pixel (for alpha) in the future.

Also introduce a new samplesPerLine property, so that tiling
is possible.